Refactor constructFormUrl func
